+++ packages/essreflectometry/src/ess/reflectometry/corrections.py
@@ -51,7 +51,7 @@ def correct_by_proton_current(da: sc.DataArray) -> sc.DataArray:
 def correct_sample_rotation(
     mu: RawSampleRotation[RunType], mu_offset: SampleRotationOffset[RunType]
 ) -> SampleRotation[RunType]:
-    return mu + mu_offset.to(unit=mu.unit)
+    return mu + mu_offset.to(unit=mu.unit, dtype=mu.dtype)
 
 
 providers = (correct_sample_rotation,)
